fix(fe): revert impl modal unification v7 (#3029)
Reverts skkuding/codedang#2964
